moved/created CreateTask component with working note creator
